spa: libcamera: source: move request completion data to `impl`
A `libcamera::Request` is directly tied to the camera, not any ports ("streams")
of it. So move the request completion ring buffer into the `impl` struct.

This is a prerequisite for supporting multiple libcamera streams (~ exposing
multiple ports on the node) in the future.

(cherry picked from commit 72fd462090971f222383e4a8d13dd6530b3f5b32)