Revert "pod: remove checks from spa_pod_body_get_*()"
This partially reverts commit f7ae61cb1ea16055bc45ec036a00ae66f66f3675.

We do want to do the checks in spa_pod_body_get_*() for extra safety.
The reason they were removed is because then we do the checks twice in
the parser. It should however be possible to fuse the can_collect and
COLLECT and SKIP calls together in the future.