stream: emit RUNNING state from Start
cd68819febdf27edfa2f8d4cda20858c5cf6c713 added code to follow the state
change of the node because at the time it the Start code was async and
it was better to complete it before emitting the state.

9b808558219f6dd37a1733c9720dcbd33178606e however made the Start command
sync again and so we can safely emit the running state in the Start.

The effect is that the Running state change is emitted first and then
the node eventfd is added to the data loop that can then call the process
callback. Having the process callback happen before the RUNNING state
change is unexpected and racy.