filter-graph: pass void * to connect_port
This makes it a bit more generic and allows us to connect other things
that float arrays.

Add a SPA_FGA_PORT_SEQUENCE as a new port type. The data to connect to
this port is a Sequence type with the size field set to the max/size of
the data.